prama-backend/requirements.txt
fastapi
uvicorn
httpx
pydantic
python-jose[cryptography]
passlib[bcrypt]
bcrypt